#642 create layouts folder and relocate nav-top-bar, page-block, page-title, and sidebar
