+++ src/frontend/pages/ChangeRequestDetailPage/change-request-details/type-specific-details/activation-details/activation-details.tsx
@@ -5,7 +5,7 @@
 
 import { ActivationChangeRequest } from 'utils';
 import { booleanPipe, datePipe, fullNamePipe } from '../../../../../../shared/pipes';
-import PageBlock from '../../../../../components/page-block/page-block';
+import PageBlock from '../../../../../layouts/page-block/page-block';
 
 interface ActivationDetailsProps {
   cr: ActivationChangeRequest;
