+++ src/frontend/pages/ChangeRequestDetailPage/change-request-details/type-specific-details/stage-gate-details/stage-gate-details.tsx
@@ -5,7 +5,7 @@
 
 import { StageGateChangeRequest } from 'utils';
 import { booleanPipe, dollarsPipe } from '../../../../../../shared/pipes';
-import PageBlock from '../../../../../components/page-block/page-block';
+import PageBlock from '../../../../../layouts/page-block/page-block';
 
 interface StageGateDetailsProps {
   cr: StageGateChangeRequest;
