+++ src/frontend/pages/CreateProjectPage/create-project-form/create-project-form.tsx
@@ -4,7 +4,7 @@
  */
 
 import { Button, Col, Form, Row } from 'react-bootstrap';
-import PageBlock from '../../../components/page-block/page-block';
+import PageBlock from '../../../layouts/page-block/page-block';
 import { CreateProjectFormStates } from '../create-project-form';
 
 interface CreateProjectFormViewProps {
