+++ src/frontend/pages/CreateWorkPackagePage/create-wp-form/create-wp-form.tsx
@@ -4,7 +4,7 @@
  */
 
 import { Form, Row, Col, InputGroup, Button } from 'react-bootstrap';
-import PageBlock from '../../../components/page-block/page-block';
+import PageBlock from '../../../layouts/page-block/page-block';
 import { EditableTextInputListUtils, FormStates } from '../create-wp-form';
 import EditableTextInputList from '../../../components/editable-text-input-list/editable-text-input-list';
