+++ src/frontend/pages/HomePage/useful-links/useful-links.tsx
@@ -5,7 +5,7 @@
  */
 
 import { linkPipe } from '../../../../shared/pipes';
-import PageBlock from '../../../components/page-block/page-block';
+import PageBlock from '../../../layouts/page-block/page-block';
 import { Col, Container, Row } from 'react-bootstrap';
 import styles from './useful-links.module.css';
