+++ src/frontend/pages/ProjectDetailPage/project-edit-container/project-edit-container.tsx
@@ -12,11 +12,11 @@ import { useAllUsers } from '../../../../services/users.hooks';
 import { useAuth } from '../../../../services/auth.hooks';
 import { EditableTextInputListUtils } from '../../CreateWorkPackagePage/create-wp-form';
 import EditableTextInputList from '../../../components/editable-text-input-list/editable-text-input-list';
-import PageTitle from '../../../components/page-title/page-title';
+import PageTitle from '../../../layouts/page-title/page-title';
 import ProjectEditDetails from './project-edit-details/project-edit-details';
 import EditModeOptions from './edit-mode-options/edit-mode-options';
 import ProjectEditSummary from './project-edit-summary/project-edit-summary';
-import PageBlock from '../../../components/page-block/page-block';
+import PageBlock from '../../../layouts/page-block/page-block';
 import ChangesList from '../../../components/changes-list/changes-list';
 import ErrorPage from '../../../pages/ErrorPage/error-page';
 import LoadingIndicator from '../../../components/loading-indicator/loading-indicator';
