+++ src/frontend/pages/ProjectDetailPage/project-view-container/project-view-container.tsx
@@ -11,8 +11,8 @@ import { useAuth } from '../../../../services/auth.hooks';
 import ChangesList from '../../../components/changes-list/changes-list';
 import DescriptionList from '../../../components/description-list/description-list';
 import WorkPackageSummary from './work-package-summary/work-package-summary';
-import PageTitle from '../../../components/page-title/page-title';
-import PageBlock from '../../../components/page-block/page-block';
+import PageTitle from '../../../layouts/page-title/page-title';
+import PageBlock from '../../../layouts/page-block/page-block';
 import ProjectDetails from './project-details/project-details';
 import RulesList from './rules-list/rules-list';
 import './project-view-container.module.css';
