+++ src/frontend/pages/SettingsPage/settings.tsx
@@ -4,8 +4,8 @@
  */
 
 import { useAuth } from '../../../services/auth.hooks';
-import PageTitle from '../../components/page-title/page-title';
-import PageBlock from '../../components/page-block/page-block';
+import PageTitle from '../../layouts/page-title/page-title';
+import PageBlock from '../../layouts/page-block/page-block';
 
 const Settings: React.FC = () => {
   const auth = useAuth();
