+++ src/frontend/pages/WorkPackageDetailPage/work-package-edit-container/work-package-edit-container.tsx
@@ -13,8 +13,8 @@ import { wbsPipe } from '../../../../shared/pipes';
 import EditableTextInputList from '../../../components/editable-text-input-list/editable-text-input-list';
 import ErrorPage from '../../../pages/ErrorPage/error-page';
 import LoadingIndicator from '../../../components/loading-indicator/loading-indicator';
-import PageBlock from '../../../components/page-block/page-block';
-import PageTitle from '../../../components/page-title/page-title';
+import PageBlock from '../../../layouts/page-block/page-block';
+import PageTitle from '../../../layouts/page-title/page-title';
 import { EditableTextInputListUtils } from '../../CreateWorkPackagePage/create-wp-form';
 import DependenciesList from '../work-package-view-container/dependencies-list/dependencies-list';
 import EditModeOptions from './edit-mode-options/edit-mode-options';
