+++ src/frontend/pages/WorkPackageDetailPage/work-package-edit-container/work-package-edit-details/work-package-edit-details.tsx
@@ -12,7 +12,7 @@ import {
   percentPipe,
   emDashPipe
 } from '../../../../../shared/pipes';
-import PageBlock from '../../../../components/page-block/page-block';
+import PageBlock from '../../../../layouts/page-block/page-block';
 
 interface Props {
   workPackage: WorkPackage;
